Changed behavior of error handling.
We include the error type in addition to the error number when reporting errors. The error type allows us to categorize errors. Also supports the output of rf_phreaker_error directly. However to support the cappeen_api, we manually change errors codes to generic errors, if they are not explicityly stated in cappeen_delegate::output_error.