Bug fix - If an error occurred during communication, sometimes we were unable to reestablish comm with the scanner without power cycling.
The bladeRF API would communicate to the FPGA, if present, when opening the unit.  If the FPGA was in a weird state this would cause a failure and we couldn't do anything without power cycling the unit.

Setting the environmental variable tells the bladeRF not to communicate to the FPGA when opening the unit.  After opening the unit we reload the FPGA and it appears to work.

We remove the variable so that in other bladeRF calls it can talk to the FPGA.