Manual frequency correction now writes it's value to the EEPROM.
Before the only freq correction value that would be written to the EEPROM was one calculated using GPS 1PPS.  However there may be cases where the freq correction value on the EEPROM becomes corrupt and the client is unable to acquire GPS, therefore manual freq correction now writes it's value to the EEPROM as well.
The unit should be warmed up for at least 5 minutes before performing manual freq correction.