Bug fix - When scanning finished collection_manager_body would cancel processing before the last packet finished processing.
Removed the extra continue msg that we emit when scanning is finished.
Implemented a counter that is equivalent to the number of items we have in flight.  We start to decrement when scanning is finished.  When it gets to zero it means all items have finished and we can cancel scanning.