Bug fix - layer 3 tracking for gsm would not reset at the proper interval.
complete_decode_interval specifies when we should start to look for layer 3 again after a successfully decode.
Because the tracker was looking at the center_freq of the measurement it would only reset that particular center_freq.
We now reset all freqs within the GSM_LOW_BANDWIDTH and GSM_HIGH_BANDWIDTH.
We may still need to account for 100khz offset...