Fix bandwidth output of lte processing
When sweeping, lte channels are added to the collection list.  Some of these
may already be in the list but with the lte channel bandwidth, not the default
bandwidth.  This gets sorted out during processing but has the side effect
of outputting a measurement with the default bandwidth.  This makes it look
like we're flipping bandwidth in the output file.

As a fix we never allow the output of lte measurements that don't match the
channel bandwidth.

We could also disallow the adding a collection_info by not including the
bandwidth in collection_info's comparison operator.