Fix power_spectrum_approximator
Use upper bound to find the bandwidth.  This ensures the scanner bandwidth
is higher than the user's span.
The minimum sampling_rate was changed so that it's high enough for the
scanner's lowest bandwidth.
Moved end frequency so that the span stays the same and the bins are
centered around the start frequency.