Add spectrum, cw, and iq
Spectrum, due to file size, is not currently logged to file.
During this upgrade we also tried to multithread collection info - it
proved to be more trouble than it was worth and is still not multithreaded.
We use the power_spectrum identifier to distinguish between different
spectrum ranges and cw.
Additional functions were added to MeasurementModel to make it easier to
get header data on the fly using roles.