Fix power spectrum binning and windowing
Reduced the min number of windows.
We make sure the scanner bandwidth is smaller than the sampling_rate.
Also we make sure that if the sampling_rate is larger than span then we
use span instead of the bandwidth closest to the sampling_rate.
Modified span to include the last bin.
When actually setting the scanner bandwidth we use the bandwidth just under
the sampling_rate.  This prevents filtering from distorting the spectrum -
not sure why this happens.