Fix usb transfer error on windows 10 after collection restart
On windows 10 a usb transfer timeout can occur when using the cypress drivers.
Bumping up the sync buffer size seems to fix the problem.  However this
does increase collection time for each packet since we have to flush the
buffer (wich is not double the size) before each snapshot transfer.